requirements.txt
# Core Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Telegram Bot
python-telegram-bot==20.7

# HTTP Clients
aiohttp==3.9.5
requests==2.31.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI and ML
openai==1.12.0
numpy>=1.26.0

# Bytez AI Integration
langchain_bytez==0.0.7
langchain==0.3.17
langchain-core>=0.3.33

# Voice Processing
elevenlabs==0.2.27

# Payment Processing
razorpay==1.4.1

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
Pillow>=10.2.0